# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tessera Reliability Explorer",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/providers/{provider_id}": {
      "put": {
        "summary": "Save Provider",
        "operationId": "save_provider_api_providers__provider_id__put",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/experiments": {
      "get": {
        "summary": "List Experiments",
        "operationId": "list_experiments_api_experiments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Experiment"
                  },
                  "type": "array",
                  "title": "Response List Experiments Api Experiments Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Start Experiment",
        "operationId": "start_experiment_api_experiments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentStarted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiments/{experiment_id}": {
      "get": {
        "summary": "Get Experiment",
        "operationId": "get_experiment_api_experiments__experiment_id__get",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Experiment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiments/{experiment_id}/resume": {
      "post": {
        "summary": "Resume Experiment",
        "operationId": "resume_experiment_api_experiments__experiment_id__resume_post",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentStarted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiments/{experiment_id}/comparisons/{variant_id}": {
      "get": {
        "summary": "Experiment Comparison",
        "operationId": "experiment_comparison_api_experiments__experiment_id__comparisons__variant_id__get",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "variant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Variant Id"
            }
          },
          {
            "name": "intervention",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "model",
                    "org",
                    "engine",
                    "grader",
                    "k",
                    "scaffold",
                    "seed",
                    "harness"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Intervention"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentComparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CategoryPairCounts": {
        "properties": {
          "matched": {
            "type": "integer",
            "title": "Matched"
          },
          "a_wins": {
            "type": "integer",
            "title": "A Wins"
          },
          "b_wins": {
            "type": "integer",
            "title": "B Wins"
          },
          "both_pass": {
            "type": "integer",
            "title": "Both Pass"
          },
          "both_fail": {
            "type": "integer",
            "title": "Both Fail"
          },
          "discordant": {
            "type": "integer",
            "title": "Discordant"
          },
          "p_value": {
            "type": "number",
            "title": "P Value"
          },
          "dropped": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dropped"
          },
          "key": {
            "type": "string",
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "matched",
          "a_wins",
          "b_wins",
          "both_pass",
          "both_fail",
          "discordant",
          "p_value",
          "dropped",
          "key"
        ],
        "title": "CategoryPairCounts"
      },
...
      "ComparisonDiagnostics": {
        "properties": {
          "a": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "type": "array",
            "title": "A"
          },
          "b": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "type": "array",
            "title": "B"
          }
        },
        "type": "object",
        "required": [
          "a",
          "b"
        ],
        "title": "ComparisonDiagnostics"
      },
...
      "Diagnostic": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "signature",
          "count"
        ],
        "title": "Diagnostic"
      },
...
      "Experiment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "done",
              "error",
              "stopped"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "baseline_variant": {
            "type": "string",
            "title": "Baseline Variant"
          },
          "request": {
            "$ref": "#/components/schemas/ExperimentRequest"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "total_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/ExperimentCell"
            },
            "type": "array",
            "title": "Cells"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at",
          "baseline_variant",
          "request",
          "error",
          "total_cost",
          "cells"
        ],
        "title": "Experiment"
      },
      "ExperimentCell": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "experiment_id": {
            "type": "string",
            "title": "Experiment Id"
          },
          "variant_id": {
            "type": "string",
            "title": "Variant Id"
          },
          "repeat_index": {
            "type": "integer",
            "title": "Repeat Index"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "done",
              "error",
              "skipped"
            ],
            "title": "Status"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "experiment_id",
          "variant_id",
          "repeat_index",
          "status",
          "run_id",
          "error"
        ],
        "title": "ExperimentCell"
      },
      "ExperimentComparison": {
        "properties": {
          "compatible": {
            "type": "boolean",
            "title": "Compatible"
          },
          "intervention": {
            "type": "string",
            "title": "Intervention"
          },
          "changed_dimensions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed Dimensions"
          },
          "unexpected_dimensions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unexpected Dimensions"
          },
          "overall": {
            "$ref": "#/components/schemas/PairCounts"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryPairCounts"
            },
            "type": "array",
            "title": "Categories"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/ComparisonDiagnostics"
          },
          "paired_repeats": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Paired Repeats"
          },
          "dropped_repeats": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Dropped Repeats"
          }
        },
        "type": "object",
        "required": [
          "compatible",
          "intervention",
          "changed_dimensions",
          "unexpected_dimensions",
          "overall",
          "categories",
          "diagnostics",
          "paired_repeats",
          "dropped_repeats"
        ],
        "title": "ExperimentComparison"
      },
      "ExperimentRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "baseline_variant": {
            "type": "string",
            "title": "Baseline Variant"
          },
          "intervention": {
            "type": "string",
            "enum": [
              "model",
              "org",
              "engine",
              "grader",
              "k",
              "scaffold",
              "seed",
              "harness"
            ],
            "title": "Intervention",
            "default": "model"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/ExperimentVariant"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 2,
            "title": "Variants"
          },
          "repeats": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Repeats",
            "default": 1
          },
          "max_cost": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Cost"
          },
          "max_consecutive_errors": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Consecutive Errors",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "name",
          "baseline_variant",
          "variants"
        ],
        "title": "ExperimentRequest"
      },
      "ExperimentStarted": {
        "properties": {
          "experiment_id": {
            "type": "string",
            "title": "Experiment Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "done",
              "error",
              "stopped"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "experiment_id",
          "status"
        ],
        "title": "ExperimentStarted"
      },
      "ExperimentVariant": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "grader": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grader"
          },
          "judge": {
            "type": "string",
            "enum": [
              "llm",
              "deterministic"
            ],
            "title": "Judge",
            "default": "deterministic"
          },
          "org": {
            "type": "string",
            "title": "Org",
            "default": "toy"
          },
          "epochs": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Epochs",
            "default": 3
          },
          "scaffold": {
            "type": "string",
            "enum": [
              "baseline",
              "refuse-aware"
            ],
            "title": "Scaffold",
            "default": "baseline"
          },
          "seed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Seed",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "model"
        ],
        "title": "ExperimentVariant"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PairCounts": {
        "properties": {
          "matched": {
            "type": "integer",
            "title": "Matched"
          },
          "a_wins": {
            "type": "integer",
            "title": "A Wins"
          },
          "b_wins": {
            "type": "integer",
            "title": "B Wins"
          },
          "both_pass": {
            "type": "integer",
            "title": "Both Pass"
          },
          "both_fail": {
            "type": "integer",
            "title": "Both Fail"
          },
          "discordant": {
            "type": "integer",
            "title": "Discordant"
          },
          "p_value": {
            "type": "number",
            "title": "P Value"
          },
          "dropped": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dropped"
          }
        },
        "type": "object",
        "required": [
          "matched",
          "a_wins",
          "b_wins",
          "both_pass",
          "both_fail",
          "discordant",
          "p_value",
          "dropped"
        ],
        "title": "PairCounts"
      },
...
      "Provider": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "readiness": {
            "type": "string",
            "enum": [
              "configured",
              "needs_config"
            ],
            "title": "Readiness"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ProviderField"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "id",
          "configured",
          "readiness",
          "fields"
        ],
        "title": "Provider"
      },
      "ProviderField": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "env_var": {
            "type": "string",
            "title": "Env Var"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          }
        },
        "type": "object",
        "required": [
          "id",
          "env_var",
          "configured"
        ],
        "title": "ProviderField"
      },
      "ProviderUpdate": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProviderUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}